components/evaluation/python/requirements.txt
# Core evaluation dependencies
langsmith>=0.1.0
python-dotenv>=1.0.0

# Optional: Development dependencies (uncomment if needed)
# pytest>=7.0.0
# black>=22.0.0
# mypy>=0.990

# Note: For production use, install from pyproject.toml:
# pip install -e .
